+++ samples/il-mflambda/b.kt
@@ -1,0 +1,2 @@
+fun runB(f: () -> Unit) { f() }
+fun main() { fromA(); runB { println("B1") } }